SVG of single-line characters for engraving
SVG containing single-line characters to be used for text on engraved panels. The characters are similar to those found in Gordon Medium, but are strictly of uniform stroke width. Character size and stroke width can be varied independently. All ASCII characters provided. 

A proper font will be created once I figure out how to do it.