AV_COOL_Panel_PCB_V2  BOM Update
BOM_V2 listed for the AV_COOL Panel in the current ECAD structure has the wrong LCSC Part number.  The original BOM file pointed towards WS2813B which is a 6 pin LED, despite the schematic calling for a WS2812B which is a 4pin.  Additionally the AV_COOL_PCB design has only 4 pads and the standard across all of the PCBs has been the 4-pin LED.

This error is more than likely a typo.  Though I've fixed it on the BOM, the origin of the error probably lies in the schematic file since the BOM is autogenerated.  Will look at the schematic later on, but for now, the adjusted BOM is correct and will be accepted by JLCPCB.

The adjusted BOM labelled "AV_COOL_PCB_V2_BOM_Fixed changes the LCSC Partnumber from
**C114592**  to  **C114586**

I've repackaged the "Manufacturing.zip" to include the fixed BOM.