Backlight_Type A PCB Manufacturing file Update
2 primary changes in this commit:

First, the naming convention was changed to remove "PCB," from the project name.  While it helps match with the OH Model, it was causing the manufacturing files to be named the same way.  JLC doesn't accept BOM's and CPL's with commas in the file name, so to keep the process smooth and quick, it's better to remove it.  Also, it was just redundant.  "PCB, [Name Panel]" located in a PCB folder which is located under the ECAD folder with everything else PCB related?  Cmon Exprezzo, you can do better.

Alright.  Second issue was the addressed and fixed was the manufacturing files.  There was something with the JLC website or algorithms that were causing a list of certain SMT components to rotate and create polarity issues.  The KiCad plugin used to create the manufacturing files recognized this and corrected this issue as it generated the BOM and CPL files.  Then JLC fixed the issue on their end, and ultimately, required the manufacturing files to be corrected to revert the rotation of the components.  (There was a list of components that were flipped, but the only two that pertained to anything OH was using was the WS2812B LEDs, Elec Cap8x10)