Power_distro_pcb Change
The "Power_Distro_PCB" has been removed from the Backlighting folder and moved to the ABSIS.  As such, the project has been named "ABSIS_PDU"...short for ABSIS_Power Distribution Unit.  While it is used in conjunction with the backlight panels, it isn't backlit itself and belongs with the ABSIS family.

Aside from moving the project, no other files have been changed.  This concludes any updates to the Backlighting_Type A folder as all PCBs belonging in that folder are backlight specfic, have been transitioned to KiCad 6, and JLC manufacturing files corrected & generated.