Completed:  ECAD ReadMe.md
Cleaned up and finished the ECAD ReadMe.  Added the installation process into the document, and by doing so, I removed the Installation.md.

The detailed "Manufacturing.md still needs to be finished and added.   But this should complete the ReadMe unless the need to separate the installation into it's own document is required.