DXF files for v0.2.0
Contains the DXF files that correspond to the parts in SVG format of the different OH structures. The parts are compiled into 5 files according to the main structures of the project: UIP, LIP, Center Tub, LC, RC. For each structure there are two files in mm and inches.

The parts have been exported from the Fusion model in mm and compiled in AUTOCAD. The inch version is obtained after a units conversion in AUTOCAD.