Mods: Guide to convert to metric (#1127)
# Description

This PR is intended to provide some basic tools and indication about how to build the OH project with metric materials.
It includes two parts: 
- A script to convert the names to the metric equivalent (from Breith)
- A guide to explain what to use instead of the intended material (from Ulukaii)

Addresses # (issue)
Metric questions

## Motivation and Context
"How can I build the project in a country that uses metric" is a common question we get. Several users already have started to use metric materials to build their pit and some projects are really advanced. This PR goal is to summarize the material choices they did and spread the knowledge directly inside the repository. This is **not** intended as an official guide to follow to convert all your pit to metric. The build is thought in imperial and this PR is not intended to change that. 

## Type of change
- PowerShell script that will change filenames
- Documentation in a README 

Please delete options that are not relevant.
- [X] Tool - New
- [X] Docs - New (A new document that is neither MCAD or ECAD)

# How Has This Been Tested?

- The script has been run several time, with all the different options, and confirmed that the names are properly changed, without typos, weird behaviour or so. Steps : Execute `Convert_Name_To_Metric.ps1` and test the different options. Git commit to the previous commit ID to get rid of the changes after testing.
- Documentation has been read several times by Ulukaii and Breith to make sure it doesn't contain typos.

* Added script to convert filenames to metric values
* More comments
* Modified to include all the informations from Ulukaii
* Moved tool to its own directory
* Update to display properly the readme on github
* Typo fix on Rowmark thickness
---------
Co-authored-by: Breith Zell <github@trimak.org>